added changes to UI and refactored asset loading
