+++ public/js/assetLoader.js
@@ -1,0 +1,36 @@
+import queue from "../../node_modules/queue-async/queue";
+import * as view from "./view";
+
+export function loader(callback){
+var meshQueue = queue();
+
+view.geometries.forEach((d) => {
+    console.log('loading ' + d[1]);
+    meshQueue.defer(function(done) {
+        view.meshLoader.load(d[1], (geom, mat) => {
+            var texture = view.textures.filter((e) => {
+                return e[0] === d[0]
+            })[0];
+            if (texture !== undefined) {
+                view.texLoader.load(texture[1], (tex) => {
+                    console.log('loading texture for ' + d[0]);
+                    view.meshes[d[0]] = new THREE.Mesh(geom, new THREE.MeshBasicMaterial({
+                        map: tex
+                    }));
+                    done(null, 'ok');
+                });
+            } else {
+                view.meshes[d[0]] = new THREE.Mesh(geom, new THREE.MeshPhongMaterial({
+                    color: 0x00ff00
+                }));
+                done(null, 'ok');
+            }
+        });
+    });
+});
+
+meshQueue.awaitAll((error, res) => {
+    console.log('assets loaded');
+    callback();
+});
+}
